+++ src/components/teach/assignments/TeachPage.scss
@@ -613,7 +613,6 @@
 
 .manage-classrooms-page .back-to-work-teach {
   .brick-row-container.teach-tab-d94 {
-    background: var(--theme-light-blue) !important;
   }
 
   .teach-assigned {
