+++ src/components/teach/assignments/components/AssignedBrickDescription.scss
