+++ src/components/teach/assignments/components/DeleteClassDialog.scss
