--- src/components/teach/assignments/components/ExpandedAssignment.tsx
@@ -1,7 +1,0 @@
-import { Assignment } from "model/classroom";
-
-export interface BookData {
-  open: boolean;
-  student: any;
-  assignment: Assignment | null;
-}
