+++ src/components/teach/assignments/components/TeachFilterSidebar.tsx
@@ -4,9 +4,9 @@ import { FormControlLabel, Grid, Radio } from "@material-ui/core";
 import "./TeachFilterSidebar.scss";
 import { ClassroomStatus, TeachClassroom } from "model/classroom";
 import SpriteIcon from "components/baseComponents/SpriteIcon";
-import EmptyFilter from "../filter/EmptyFilter";
+import EmptyFilter from "./EmptyFilter";
 import RadioButton from "components/baseComponents/buttons/RadioButton";
-import CreateClassDialog from "components/teach/manageClassrooms/components/CreateClassDialog";
+import CreateClassDialog from "./CreateClassDialog";
 import { User } from "model/user";
 import SortButton from "./SortButton";
 import { Subject } from "model/brick";
